Implement `_parse_config` method
We put the logic for parsing hierarchy config into a separate method to
lower the level of complexity for `get_vars` method.

The method put two parameters `entity` the entity it works on and
optional `parse` to define for what kind of entities the method should
return hierarchy. Default for `parse` is `both`.