Finish first working version of hieradata vars plugin
We removed the distinction of entity type of host and group. We only
support host vars because we don't like that concept of group_vars so
much. With this plugin we provide a proper tool for defining variables
for different scenarios based on host name. This is a valid way to work
on, because we know of the existence of host naming schemes to describe
a lot of each machine.
We finally implement the way to load variables from files within
hierarchy.
Last but not least we add some example data to test the behavior of our
plugin. First tests are looking quiet good.