Switch to wrapper for combine_vars
As we want to provide in later versions better and extensive merge
options and as the `hash_behaviour` option is deprecated in ansible we
decided to provide a wrapper function for `combine_vars`. This function
now supports `hash_behavior` and `list_behavior` our understanding of
configuring a merger.
For that we add these options to our `ansible.cfg` section as well and
load these vars in `get_vars` as well.